+++ src/common.ts
@@ -1,3 +1,5 @@
+import { XmlPrimitiveType, XmlType } from './types';
+
 export function errUnknownClass(classConstructor: any): Error {
   return new Error(
     `Class "${classConstructor}" not found. Make sure there is a @XmlEntity({...}) decorator on it, or @XmlProperty({...}) decorator on its properties.`,
@@ -9,9 +11,17 @@ export function serializeUnionForLog(union: any[]): string {
     '[' +
     union
       .map((t) =>
-        t === null ? 'null' : t === undefined ? 'undefined' : t.name,
+        t === null
+          ? 'null'
+          : t === undefined
+          ? 'undefined'
+          : t.name || `${union}`,
       )
       .join(', ') +
     ']'
   );
+}
+
+export function isPrimitiveType(type: XmlType): type is XmlPrimitiveType {
+  return type === String || type === Number || type === Boolean;
 }