+++ types/index.d.ts
@@ -1,5 +1,5 @@
 export { classToXml } from './transform-class-to-xml';
 export { xmlToClass } from './transform-xml-to-class';
 export { XmlEntity, XmlProperty, XmlAttribute } from './decorators';
-export type { AnyClass, ClassToXmlOptions, XmlEntityOptions, XmlPropertyOptions, XmlType, } from './types';
+export type { XmlClass, ClassToXmlOptions, XmlEntityOptions, XmlPropertyOptions, XmlType, } from './types';
 //# sourceMappingURL=index.d.ts.map