+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## WIP v2.0.0
+
+- support circular dependencies by specifying `type` and `union` only via functions
+- rename `AnyClass` to `XmlClass` and add comments about constructor arguments
+- internal: fix `serializeUnionForLog` error messages
+- internal: add `isPrimitiveType` helper
+
 ## v1.0.3
 
 [Compare](https://github.com/edgar-p-yan/xml-class-transformer/compare/v1.0.2...v1.0.3)